grpc-py-client/requirements.txt
google==3.0.0
grpcio==1.54.2
grpcio-tools==1.54.2
